MBL-13231: Cleaned up
